add crud client

+++ src/main/java/com/projet_micro_service/client/controller/ClientController.java
@@ -3,10 +3,13 @@ package com.projet_micro_service.client.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
@@ -35,19 +38,38 @@ public class ClientController {
     }
 
     @GetMapping("/clients/{id}")
-    public ClientModel getClientById(@PathVariable int id) {
-        return clientDao.findById(id);
+    public ResponseEntity<ClientModel> getClientById(@PathVariable int id) {
+        ClientModel client = clientDao.findById(id);
+        if (client != null) {
+            return ResponseEntity.ok(client);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping("/clients")
-    public ClientModel createClient(@RequestBody ClientModel client) {
-        return clientDao.save(client);
+    public ResponseEntity<ClientModel> createClient(@RequestBody ClientModel client) {
+        ClientModel savedClient = clientDao.save(client);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedClient);
+    }
+
+    @PutMapping("/clients/{id}")
+    public ResponseEntity<ClientModel> updateClient(@PathVariable int id, @RequestBody ClientModel client) {
+        if (!clientDao.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+        client.setId(id);
+        ClientModel updatedClient = clientDao.save(client);
+        return ResponseEntity.ok(updatedClient);
     }
 
     @DeleteMapping("/clients/{id}")
-    public String deleteClient(@PathVariable int id) {
+    public ResponseEntity<String> deleteClient(@PathVariable int id) {
+        if (!clientDao.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         clientDao.deleteById(id);
-        return "{\"message\":\"Client with id " + id + " deleted\", \"status\":200}";
+        return ResponseEntity.ok("{\"message\":\"Client with id " + id + " deleted\", \"status\":200}");
     }
 
     @GetMapping("/clients/{id}/orders")
